Funcionalidades comprobadas y modificación de componentes xhtml.
Se ha modificado footer.xhtml, index.html y template.xhmtl

Se ha modificado main.css y añadido home.css (pendiente de revisión para factorización).

Se ha comprobado el correcto funcionamiento de las nuevas funcionalidades de la entrega 2.

+++ src/main/java/dominio/LugarDeRecogida.java
@@ -25,7 +25,7 @@ public class LugarDeRecogida {
 	public LugarDeRecogida() {};
 	
 	
-	public LugarDeRecogida(String recogida, int longitud, int latitud) {
+	public LugarDeRecogida(String recogida, double longitud, double latitud) {
 		this.recogida = recogida;
 		this.longitud = longitud;
 		this.latitud = latitud;
